@@ -27,8 +27,8 @@
     "lodash-es": "^4.17.21",
     "mdi-material-ui": "^7.4.0",
     "notistack": "^2.0.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-error-boundary": "^3.1.4",
     "react-query": "^3.39.2",
     "use-immer": "^0.7.0",
